+++ mappings.py
@@ -1,0 +1,8 @@
+pay_cycle_map = {
+    "weekly": 52,
+    "bi_weekly": 26,
+    "semi_monthly": 24,
+    "monthly": 12,
+    "other": None           # this will bite me later if I don't solve what other means. Probably don't include 'other', just split it into other choices that will appear here.
+}
+